backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
psycopg2-binary>=2.9.6
python-dotenv>=1.0.0
Pillow>=9.5.0
django-filter>=23.2
pytest>=7.3.1
pytest-django>=4.5.2
